feat(frontend): removed fixed modal dimensions
The dimensions were rolled back because the modal will eventually
implement a grid wrap for each form row. Besides, there are plans
to not use modals in the future, instead, use another emerging
design pattern.

The forms rows need to be in a uniform grid. The idea is to implement
the grid on all tabs and it will force the modal into a uniform width,
instead of doing manual hacks. This will also fix responsiveness because
right now all responsive stuff implemented is broken.